% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicAlluvialPlot.R
\name{BasicAlluvialPlot}
\alias{BasicAlluvialPlot}
\title{BasicAlluvialPlot}
\usage{
BasicAlluvialPlot(data)
}
\arguments{
\item{data}{A \code{data.frame} object in the format produced by  \code{BarcodeFlow::FormatData}}
}
\value{
A \code{ggplot2} object.
}
\description{
Takes in a formatted table and produces a basic alluvial
}
\details{
\code{data.frame} object is expected to be in the format produced by \code{BarcodeFlow::FormatData}. But does not technically require it to work. Can have additional columns.
}
\examples{
BasicAlluvialPlot(df_barcodes)

}
