% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteCombinations.R
\name{CompleteCombinations}
\alias{CompleteCombinations}
\title{CompleteCombinations}
\usage{
CompleteCombinations(data)
}
\arguments{
\item{data}{The \code{data.frame} object to have observations completed.}
}
\value{
A \code{data.frame} object with complete observations for each sample_name_id, variable, and time_point.
}
\description{
Creates complete observations for each each sample_name_id, variable, and time_point column in \code{data}.
}
\details{
Helper function for \code{BarcodeFlow::FormatData}. Assumes column names are 'sample_name_id', 'variable', 'time_point', 'variable_proportion'.
}
\examples{
CompleteCombinations(df_barcodes)

}
