% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_barcodes.R
\docType{data}
\name{df_barcodes}
\alias{df_barcodes}
\title{Dataset of randomly created barcode counts.}
\format{
\subsection{\code{df_barcodes}}{

A data frame with 400 rows and 7 columns:
\describe{
\item{barcode}{12 character ACTG barcode}
\item{sample_name}{sample name given}
\item{animal_id}{fake name of animal}
\item{dpi}{fake days post infection assigned}
\item{celltype}{fake cell type or stock}
\item{count}{randomly sampled count for a barcode}
\item{proportion}{calculated per sample_name and adds up to 1}
}
}
}
\usage{
df_barcodes
}
\description{
Barcode counts were obtained by sampling from a poisson distribution of
decreasing lambda values for increase DPI values to mimic selection.
}
\keyword{datasets}
