% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatData.R
\name{FormatData}
\alias{FormatData}
\title{FormatData}
\usage{
FormatData(data, sample_name_id, variable, time_point, variable_proportion)
}
\arguments{
\item{data}{The \code{data.frame} object to be used for plotting.}

\item{sample_name_id}{The column name containing a sample identifier.}

\item{variable}{The column name containing the variable that is being examined. Must be in the same unit i.e. Days/Months/Years}

\item{time_point}{The column name containing time point information.}

\item{variable_proportion}{The column name that has the proportion value for each variable that was measured.}
}
\value{
A \code{data.frame} object with standardized column name outputs.
}
\description{
Extract relevant features from a dataframe necessary for alluvial plotting.
}
\details{
Generates a \code{data.frame} object with original data and additional columns that are ordered factors to be used for plotting.
Uses \code{BarcodeFlow::CompleteCombinations()}
}
\examples{
FormatData(df_barcodes, 'sample_name', 'barcode', 'dpi', 'proportion')

}
